Renaming of interface files
