+++ src/app/events.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
 // import { Event } from './interfaces/event'
-import { User } from './interfaces/user'
+import { User } from './interfaces/user.interface'
 
 @Injectable({
   providedIn: 'root'
