+++ src/app/home-page/home-page.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from "@angular/core";
 import { Observable } from "rxjs";
 
-import { User } from "../interfaces/user";
+import { User } from "../interfaces/user.interface";
 import { EventsService } from "../events.service";
 
 @Component({
