+++ src/app/interfaces/comment.interface.ts
@@ -1,0 +1,8 @@
+import { User } from "./user.interface";
+import { Event } from "./event.interface";
+export class Comment {
+  event: Event;
+  author: User;
+  comment_text: string;
+  pub_date: Date;
+}
