+++ src/app/interfaces/event.interface.ts
